Tools: Include Mbed config header in ARMC6 linker preprocessing
Currently, when building with ARMC6 we check at the linking stage for
the presence of a shebang line, compare this to the shebang generated by
the build system and replace it if it is different.

We alter this to always change the shebang, if one is present, in order
to add the inclusion of the Mbed config header. If the scatter file is
preprocessed, we need to do so in a context where the Mbed configuration
defines are set correctly. If no shebang line is present, no
preprocessing is occuring and we just return the unchanged scatter file
as before.