Connectivity: Make open(NetworkStack*) virtual in class Socket
The method `nsapi_error_t open(NetworkStack *stack)` is implemented for
all Concrete classes derived from the Socket Abstract Base Class,
excluding CellularNonIPSocket. It is defined seperately in
InternetSocket and TLSSocketWrapper as a non virtual function, which
means that it cannot be readily Mocked during unit testing.

In order to solve this, we move the method into Socket and mark it as
not implemented in CellularNonIPSocket.